feat: 1前端使用vue-cropper裁剪图片 2增加注册邮箱验证，登录验证码，图片存储格式一律使用webp，极大减少图片空间 3前端布局，后端管理界面增加话题管理

+++ acimage_image/src/main/java/com/acimage/image/service/imagehash/impl/ImageHashWriteServiceImpl.java
@@ -1,5 +1,6 @@
 package com.acimage.image.service.imagehash.impl;
 
+import cn.hutool.core.collection.CollectionUtil;
 import com.acimage.image.dao.ImageHashDao;
 import com.acimage.image.model.domain.ImageHash;
 import com.acimage.image.service.imagehash.ImageHashWriteService;
@@ -22,7 +23,10 @@ public class ImageHashWriteServiceImpl implements ImageHashWriteService {
 
     @Override
     public void removeImageHashes(List<Long> imageIds) {
-            imageHashDao.deleteBatchIds(imageIds);
+        if (CollectionUtil.isEmpty(imageIds)) {
+            return;
+        }
+        imageHashDao.deleteBatchIds(imageIds);
     }
 
     @Override
